refactor: 메서드 조건분기 간소화(다른 클래스에 기능 위임)
