+++ src/test/java/lotto/domain/LottoGameTest.java
@@ -4,13 +4,11 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.Arrays;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 public class LottoGameTest {
 
@@ -24,20 +22,6 @@ public class LottoGameTest {
         winningLottoExcludingBonus = new Lotto(Arrays.asList(1, 2, 3, 4, 5, 45));
         bonusLottoNumber = new LottoNumber(6);
         winningLotto = new WinningLotto(winningLottoExcludingBonus, bonusLottoNumber);
-    }
-
-    @ParameterizedTest
-    @CsvSource(value = {"3:5000", "4:50000", "5:1500000", "6:2000000000"}, delimiter = ':')
-    void 수익_계산(int matchCount, long reward) {
-        assertThat(LottoGame.reward(matchCount)).isEqualTo(reward);
-    }
-
-    @ParameterizedTest
-    @ValueSource(ints = {1, 2, 7})
-    void 번호_일치_개수_예외(int matchCount) {
-        assertThatIllegalArgumentException()
-                .isThrownBy(() -> LottoGame.reward(matchCount))
-                .withMessageContaining("당첨에 해당하는 번호 일치 개수가 아닙니다.");
     }
 
     @ParameterizedTest
@@ -80,8 +64,8 @@ public class LottoGameTest {
         LottoRewards reward = LottoGame.reward(lottos, winningLotto);
 
         //then
-        assertThat(reward.get(RewardType.FIVE_AND_BONUS).count()).isEqualTo(1);
-        assertThat(reward.get(RewardType.SIX).count()).isEqualTo(1);
+        assertThat(reward.get(RewardType.FIVE_AND_BONUS)).isNotNull();
+        assertThat(reward.get(RewardType.SIX).count()).isNotNull();
     }
 
     @Test
@@ -93,6 +77,7 @@ public class LottoGameTest {
         LottoRewards reward = LottoGame.reward(lottos, winningLotto);
 
         //then
-        assertThat(reward.get(RewardType.FIVE_AND_BONUS).count()).isEqualTo(1);
+        assertThat(reward.get(RewardType.FIVE_AND_BONUS)).isNotNull();
+        assertThat(reward.get(RewardType.FIVE)).isNull();
     }
 }
