+++ src/test/java/lotto/domain/RewardTypeTest.java
@@ -1,0 +1,16 @@
+package lotto.domain;
+
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
+
+import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
+
+public class RewardTypeTest {
+    @ParameterizedTest
+    @ValueSource(ints = {1, 2, 7})
+    void 번호_일치_개수_예외(int matchCount) {
+        assertThatIllegalArgumentException()
+                .isThrownBy(() -> RewardType.of(matchCount, false))
+                .withMessageContaining("당첨에 해당하는 번호 일치 개수가 아닙니다.");
+    }
+}
